+++ src/FluxFlow.Components.Mapping/Contracts/IMappingContextFactory.cs
@@ -1,0 +1,8 @@
+using FluxFlow.Engine.Mapping;
+
+namespace FluxFlow.Components.Mapping.Contracts;
+
+public interface IMappingContextFactory
+{
+    FlowMapContext Create(object? input, MappingNodeContext context);
+}