+++ src/FluxFlow.Components.Mapping/MappingComponentTypes.cs
@@ -1,0 +1,8 @@
+using FluxFlow.Engine.Definitions;
+
+namespace FluxFlow.Components.Mapping;
+
+public static class MappingComponentTypes
+{
+    public static readonly NodeType Mapper = new("flow.mapper");
+}