+++ src/FluxFlow.Components.Mapping/MappingErrorCodes.cs
@@ -1,0 +1,6 @@
+namespace FluxFlow.Components.Mapping;
+
+public static class MappingErrorCodes
+{
+    public const int MapperFailed = 3000;
+}